feat(months): add month and day components which properly display each week in a given month
Almost burst a vein making sure the days would shift according to the specified rules. It was a matter of how I was looking at the problem. I was trying to generate the "day array" for the month with the leading and trailing "empty days". The solution ended up being to simply generated the "empty" days after having generated the regular "day array".

closes #2